silence warnings for ILP64
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@5242 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_zgegqr_gpu.cpp
@@ -50,7 +50,7 @@ int main( int argc, char** argv)
     
     // versions 1...4 are valid
     if (opts.version < 1 || opts.version > 4) {
-        printf("Unknown version %d; exiting\n", opts.version );
+        printf("Unknown version %d; exiting\n", (int) opts.version );
         return -1;
     }
